@@ -10,24 +10,24 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.2.4",
-    "@angular/common": "~13.2.4",
-    "@angular/compiler": "~13.2.4",
-    "@angular/core": "~13.2.4",
-    "@angular/forms": "~13.2.4",
-    "@angular/localize": "~13.2.4",
-    "@angular/platform-browser": "~13.2.4",
-    "@angular/platform-browser-dynamic": "~13.2.4",
-    "@angular/router": "~13.2.4",
+    "@angular/animations": "^14.0.1",
+    "@angular/common": "^14.0.1",
+    "@angular/compiler": "^14.0.1",
+    "@angular/core": "^14.0.1",
+    "@angular/forms": "^14.0.1",
+    "@angular/localize": "^14.0.1",
+    "@angular/platform-browser": "^14.0.1",
+    "@angular/platform-browser-dynamic": "^14.0.1",
+    "@angular/router": "^14.0.1",
     "@fortawesome/angular-fontawesome": "^0.10.1",
     "@fortawesome/fontawesome-svg-core": "^1.3.0",
-    "@fortawesome/free-solid-svg-icons": "^6.0",
     "@fortawesome/free-regular-svg-icons": "^6.0",
+    "@fortawesome/free-solid-svg-icons": "^6.0",
     "@ng-bootstrap/ng-bootstrap": "^12.0.0",
     "@ngx-translate/core": "^14.0.0",
     "@ngx-translate/http-loader": "^7.0.0",
     "@popperjs/core": "^2.11.2",
-    "bootstrap": "^5.1.3",
+    "bootstrap": "^5.2.0-beta1",
     "ngx-logger": "^5.0.8",
     "ngx-spinner": "^13.1.1",
     "ngx-toastr": "^14.2.1",
@@ -36,9 +36,9 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.2.5",
-    "@angular/cli": "^13.2.5",
-    "@angular/compiler-cli": "~13.2.4",
+    "@angular-devkit/build-angular": "^14.0.1",
+    "@angular/cli": "^14.0.1",
+    "@angular/compiler-cli": "^14.0.1",
     "@types/jasmine": "~3.10.3",
     "@types/node": "^17.0.21",
     "jasmine-core": "~4.0.1",
@@ -48,6 +48,6 @@
     "karma-jasmine": "~4.0.1",
     "karma-jasmine-html-reporter": "^1.7.0",
     "prettier": "2.5.1",
-    "typescript": "~4.5.5"
+    "typescript": "~4.7.3"
   }
 }
